+++ CHANGES.rst
@@ -13,6 +13,13 @@ Unreleased
 
 * Update entries whenever their :attr:`~Entry.updated` changes,
   don't compare the values. (:issue:`332`)
+* Add :attr:`Feed.update_after` and :attr:`Feed.last_retrieved`. (:issue:`332`)
+* The ``new`` filter of :meth:`~Reader.update_feeds()` etc. considers
+  a feed as new if it was never retrieved (:attr:`~Feed.last_retrieved`),
+  not if it was never updated successfully (:attr:`~Feed.last_updated`). (:issue:`332`)
+
+  .. FIXME: versionchanged on update_feeds() etc.
+
 * Group mutually-exclusive attributes of :class:`~.FeedUpdateIntent`
   into its :attr:`~.FeedUpdateIntent.value` union attribute. (:issue:`332`)